removing yellow warnings

+++ AC3.2-Cheers/Venue.swift
@@ -60,7 +60,7 @@ class Venue {
         return venues
     }
     
-    func convert(meters: Int16) -> String {
+    func convert(meters: Int) -> String {
         let miles = Double(meters) * 0.0006
         return String(format: "%.1f", miles) + " miles away"
     }
